Structural inital refactor
